gen/templates improvements (#62)
The `go fmt` always puts an empty line after a function and before the `// End` comment. Do the same.

There are tests generated now for any use of `data_source_name_query`.

In model.go: keep the behavior but swap the branches. It's more readable when a special case (`break`) is handled inside an `if` block and normal case outside of `if` block.

Don't mix tabs/spaces in indentation.