fix: in Create and Update never attempt to modify Known values (#86)
fix: Create/Update methods only modify value if IsUnknown

Fix of a slight bug exposure. The existing `updateFromBody` might sometimes read a Known state value during Create or Update (whenever `target.IsNull()`). In this situation terraform complains loudly that this is "always a bug in provider", because we break the contract of Create or Update.

After this change, Create or Update only read unknown values. They use the heuristic that unknowns will be given only on `Computed` attributes, but always check at runtime with IsUnknown(). So that heuristic is just for simplifying code.

Also, rename the `updateFromBody()` to `fromBodyPartial()`, because it is no longer called from within Update. Keeping the old name would be probably misleading.