templates: reduce repeated code, add Map (#118)
# Recursion in Templates

Reduce the repeated occurrences of nearly-identical code. The reason is I tried to add a major functionality to templates and my progress was grinding nearly to a halt. Subjectively, I was having major problems even coherently editing `gen/templates/model.go`. The only way to proceed was to nest subtemplates recursively:

     - considerably reduces template code (the code reduced might be even untested/dead code: no yaml currently generates one of the levels)
     - the remaining code is left more readable
     - code with subtemplates is infinitely nestable (depth not capped at 4)

# Nested Map

A nested Map will be useful for functionality like bulk network groups (resource fmc_network_groups) or other bulk objects.

Looking in general, nested Map is marshaled into JSON array to be compatible with a FMC bulk-POST, for example:

```
[
  {"name":"...","id":"...","attr":"..."},
  {"name":"...","id":"...","attr":"..."}
]
```

It might be surprising, since natural marshaling of a Map is a JSON object (instead of a JSON array). Presently FMC API doesn't however need a JSON **object** with arbitrary keys and we still can do that in future.

We don't want to use List here, because after the first `terraform apply` the other Terraform resources depend on the id and need to be ready for the subsequent run of `terraform apply`:

```
  my_object_id = some_list[42].id      # UNSTABLE when removing element 0-41
```

The subsequent runs become more stable with Map:

```
  my_object_id = some_map["name"].id     # stable even when other elements are later removed/added
```

# Yaml Validation

Add more yaml validation - for Map and for the existing List/Set, fail hard when user specifies nonsensical/unsupported yaml.